% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_microbiome.R
\name{extract_microbiome}
\alias{extract_microbiome}
\title{Extract Microbiome reads and Kraken output}
\usage{
extract_microbiome(
  fq1,
  fq2 = NULL,
  kraken_out = NULL,
  kraken_report = NULL,
  mpa_report = NULL,
  out_dir = getwd(),
  sample = NULL,
  microbiome_pattern = "(?i)Bacteria|Fungi|Viruses",
  ntaxid = 8000L,
  ...,
  sys_args = list()
)
}
\arguments{
\item{fq1, fq2}{The path to the classified microbiome fastq file (returned by
\code{run_kraken2}).}

\item{kraken_out}{The path to kraken output file if \code{NULL}, will locate
\strong{sample.kraken.output.txt} in \code{out_dir}.}

\item{kraken_report}{The path to kraken2uniq report if \code{NULL},
\strong{sample.kraken.report.txt} in \code{out_dir}.}

\item{mpa_report}{The path to mpa style report if \code{NULL},
\strong{sample.kraken.report.mpa.txt} in \code{out_dir}.}

\item{out_dir}{The path of output directory.}

\item{sample}{A string, sample name, will be used to create output file name.}

\item{microbiome_pattern}{Perl regex patterns to match the bacteria
microbiome.}

\item{ntaxid}{Number of taxids to extract at a time.}

\item{...}{Other arguments passed to \link{grepl}.}

\item{sys_args}{Other arguments passed to \link{system2}.}
}
\description{
Extract Microbiome reads and Kraken output
}
\seealso{
\url{https://github.com/sjdlabgroup/SAHMI}
}
