% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koutreads.R
\name{koutreads}
\alias{koutreads}
\alias{tag}
\title{Kraken2 Output Processing for FASTQ Files}
\usage{
koutreads(
  kreport,
  koutput,
  reads,
  tag_ranges1 = NULL,
  tag_ranges2 = NULL,
  ofile = NULL,
  taxonomy = c("d__Bacteria", "d__Fungi", "d__Viruses"),
  exclude = c("9606"),
  koutput_batch = NULL,
  fastq_batch = NULL,
  chunk_bytes = NULL,
  compression_level = 4L,
  nqueue = NULL,
  threads = NULL,
  odir = NULL
)

tag(tag, ranges)
}
\arguments{
\item{kreport}{Path to the Kraken2 report file.}

\item{koutput}{Path to the Kraken2 output file.}

\item{reads}{A character vector of FASTQ file paths, either the original
reads used as input to Kraken2 or the classified output reads (recommended
for efficiency as they are smaller). Accepts one file for single-end or two
files for paired-end.

\strong{If only one file is used in Kraken2 to generate the koutput file, only the
second read sequence will be extracted to match the koutput's Lowest Common
Ancestor (LCA)}. This is commonly observed in data such as 10x Genomics,
where the first read contains only UMI and barcode information, while the
second read carries the biological sequence. You can always change the input
files to treat the actual biological read as read1 if needed.}

\item{tag_ranges1, tag_ranges2}{A list of sequence ranges for extracting tags
from the first/second read in single-end/paired-end data. If \code{NULL}, no
additional tag extraction occurs for the first/second read. These ranges
must be created by the \code{tag()} function. Note that the tag embedded by
\code{\link[=seq_refine]{seq_refine()}} in the description header will always be extracted.}

\item{taxonomy}{A character vector. The set of taxonomic groups to include
(default: \code{c("d__Bacteria", "d__Fungi", "d__Viruses")}). This defines the
global taxa to consider. Only the descendants within these groups will be
considered. If \code{NULL}, all taxa will be used.}

\item{exclude}{A character vector of taxids to exclude sequences from usage.
Typically used to exclude the host taxid (e.g., \code{9606} for human) from the
analysis. By default, this excludes human sequences (\code{"9606"}).}

\item{koutput_batch, fastq_batch}{Integer. Number of FASTQ records/Koutput
lines to accumulate before dispatching a chunk to worker threads for
processing. This controls the granularity of parallel work and affects
memory usage and performance.
Default is \code{1000} for \code{koutput_batch}
and \code{256} for \code{fastq_batch}.}

\item{chunk_bytes}{Integer specifying the size in bytes used for compressing
and writing records in batches to disk. Default is \code{8 * 1024 * 1024}
(8MB).}

\item{compression_level}{Integer from 1 to 12 (default: \code{4}). This sets the
gzip compression level when writing output files. A higher value increases
compression ratio but may slow down writing. Only applies when output
filenames end with \code{.gz}.}

\item{nqueue}{Integer. Maximum number of buffers per thread, controlling the
amount of in-flight data awaiting writing. Default: \code{3}.}

\item{threads}{Integer. Number of threads to use. Default: \code{3}.}

\item{odir}{A string of directory to save the output files. Please see
\code{Value} section for details.}

\item{tag}{An character label used to label the extracted content.}

\item{ranges}{A range or a list of ranges specifying the subsequence(s) to
process. Must be created using the \code{\link[=seq_range]{seq_range()}} function.}
}
\description{
This function processes Kraken2 output and associated FASTQ files, extracting
relevant taxonomic information and preparing them for further downstream
analysis.
}
