% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kractor.R
\name{kractor_koutput}
\alias{kractor_koutput}
\title{Filter Kraken2 Output by Taxon}
\usage{
kractor_koutput(
  kreport,
  koutput,
  ofile = NULL,
  taxonomy = c("D__Bacteria", "D__Fungi", "D__Viruses"),
  ranks = NULL,
  taxa = NULL,
  taxids = NULL,
  exclude = NULL,
  descendants = TRUE,
  batch_size = NULL,
  chunk_bytes = NULL,
  compression_level = 4L,
  nqueue = NULL,
  threads = NULL,
  odir = NULL
)
}
\arguments{
\item{kreport}{Path to the Kraken2 report file.}

\item{koutput}{Path to the Kraken2 output file.}

\item{taxonomy}{Character vector. The set of taxonomic groups to include
(default: \code{c("D__Bacteria", "D__Fungi", "D__Viruses")}). This defines the
global taxa to consider. If \code{NULL}, all taxa will be used. If \code{descendants = TRUE}, only the descendants within these groups will be considered. The
selection of taxa can be further refined using the \code{ranks}, \code{taxa}, and
\code{taxids} parameters. One of \code{taxonomy}, \code{ranks}, \code{taxa}, or \code{taxids} must be
provided.}

\item{ranks}{Character vector. The taxonomic ranks to filter by (optional).}

\item{taxa}{Character vector. Specific taxa to include (optional).}

\item{taxids}{Character vector. A list of taxid values to filter by
(optional).}

\item{exclude}{A character vector of taxids to exclude sequences from usage.}

\item{descendants}{Logical. Whether to include descendants of the selected
taxa (default: \code{TRUE}).}

\item{chunk_bytes}{Integer specifying the size in bytes used for compressing
and writing records in batches to disk. Default is \code{8 * 1024 * 1024}
(8MB).}

\item{compression_level}{Integer from 1 to 12 (default: \code{4}). This sets the
gzip compression level when writing output files. A higher value increases
compression ratio but may slow down writing. Only applies when output
filenames end with \code{.gz}.}

\item{nqueue}{Integer. Maximum number of buffers per thread, controlling the
amount of in-flight data awaiting writing. Default: \code{3}. Setting this too
high may increase memory consumption without performance gain.}

\item{threads}{Integer. Number of threads to use. Default: \code{3}.}

\item{odir}{A string of directory to save the output files. Please see
\code{Value} section for details.}
}
\value{
None. The function generates a filtered Kraken2 output file
containing entries corresponding to the specified \code{taxonomy}, \code{ranks},
\code{taxa}, \code{taxids}, and \code{descendants} extracted from the input \code{koutput}.
}
\description{
This function filters Kraken2 classification output (\code{koutput}) by taxonomic
identifiers derived from a Kraken2 report (\code{kreport}). It extracts lines
matching the desired \code{taxonomy}, \code{ranks}, \code{taxa}, \code{taxids}, and \code{descendants}
and writes the filtered results to an output file.
}
