% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken2.R
\name{kraken2}
\alias{kraken2}
\title{Run Kraken2 for Taxonomic Classification}
\usage{
kraken2(
  reads,
  ...,
  db = NULL,
  kreport = "kraken_report.txt",
  koutput = "kraken_output.txt",
  classified_out = "classified.fq",
  unclassified_out = NULL,
  kraken2 = NULL,
  envpath = NULL,
  conda = NULL,
  condaroot = NULL,
  odir = NULL
)
}
\arguments{
\item{reads}{A character vector of FASTQ files used as input to Kraken2.
Can be one file (single-end) or two files (paired-end).}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \code{kraken2} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link[=shQuote]{shQuote()}}. Details see: \code{cmd_help(kraken2())}.}

\item{db}{Path to the Kraken2 database. You can download prebuilt databases
from \url{https://benlangmead.github.io/aws-indexes/k2}, or build your own by
following the instructions at
\url{https://github.com/DerrickWood/kraken2/wiki/Manual#kraken-2-databases}.}

\item{classified_out}{A string of path to save classified sequences, which
should be a fastq file.}

\item{unclassified_out}{A string of path to save unclassified sequences,
which should be a fastq file.}

\item{kraken2}{Optional. Path to the Kraken2 binary if not in the system
\code{PATH}.}

\item{envpath}{Optional. Additional directory to prepend to the system \code{PATH}
environment variable.  Useful when \code{kraken2} is installed in a non-default
location.}

\item{conda}{Optional. Name of the Conda environment to activate before
running Kraken2.}

\item{condaroot}{A string specifying the path to the conda root prefix. If
not provided, the function searches for the root in the following order:
\enumerate{
\item the \link{option} \code{blit.conda.root}.
\item the \link[=Sys.getenv]{environment variable} \code{BLIT_CONDA_ROOT}.
\item the root prefix of \code{\link[blit:appmamba]{appmamba()}}.
}}

\item{odir}{A string of path to the output directory.}
}
\value{
None. This function is called for its side effects.
It produces the following output files in \code{odir} (or the working directory
if \code{odir} is \code{NULL}):
\itemize{
\item \code{kreport}: Kraken2 classification report
\item \code{koutput}: Kraken2 raw classification output
\item \code{classified_out}: FASTQ file of classified reads
\item \code{unclassified_out}: FASTQ file of unclassified reads (if specified)
}
}
\description{
This function runs the Kraken2 classifier on one or two FASTQ files (single-
or paired-end). It wraps around the \code{blit::kraken2()} command interface and
adds optional support for environment setup (e.g., Conda or custom \code{PATH}).
}
