% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken_microbiome.R
\name{kraken2_microbiome}
\alias{kraken2_microbiome}
\title{Extracting microbiome kraken2 output and sequence reads}
\usage{
kraken2_microbiome(
  fq1,
  ...,
  fq2 = NULL,
  db = NULL,
  taxon = c("d__Bacteria", "d__Fungi", "d__Viruses"),
  kreport = "kraken_report.txt",
  koutput = "kraken_output.txt",
  classified_out = "classified.fq",
  unclassified_out = NULL,
  microbiome_koutput = "kraken_microbiome_output.txt",
  microbiome_reads = NULL,
  threads = NULL,
  odir = getwd(),
  kraken2 = NULL,
  envpath = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{fq1, fq2}{A string of fastq file path.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \code{kraken2} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link[=shQuote]{shQuote()}}. Details see: \code{cmd_help(kraken2())}.}

\item{taxon}{An atomic character specify the taxa name wanted. Should follow
the kraken style, connected by rank codes, two underscores, and the
scientific name of the taxon (e.g., "d__Viruses")}

\item{classified_out}{A string of path to save classified sequences, which
should be a fastq file.}

\item{unclassified_out}{A string of path to save unclassified sequences,
which should be a fastq file.}

\item{threads}{Number of threads to use, see
\code{blit::cmd_help(blit::seqkit("grep"))}.}

\item{odir}{A string of path to the output directory.}

\item{kraken2}{A string of path to \code{kraken2} command.}

\item{envpath}{A string of path to be added to the environment variable
\code{PATH}.}
}
\description{
Extracting microbiome kraken2 output and sequence reads
}
