% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken_report.R
\name{kraken_report}
\alias{kraken_report}
\title{Parse kraken report file}
\usage{
kraken_report(kreport)
}
\arguments{
\item{kreport}{The path to kraken report file.}
}
\value{
A data frame.
}
\description{
Parse kraken report file
}
\seealso{
\url{https://github.com/DerrickWood/kraken2/blob/master/docs/MANUAL.markdown}
}
