% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krakenx.R
\name{krakenx}
\alias{krakenx}
\title{Full SAHMI-Compatible Pipeline: Kraken2 + Read Extraction + K-mer/UMI
Quantification}
\usage{
krakenx(
  reads,
  ...,
  ubread = NULL,
  umi_ranges = NULL,
  barcode_ranges = NULL,
  barcode_extractor = NULL,
  umi_extractor = NULL,
  db = NULL,
  kreport = "kraken_report.txt",
  koutput = "kraken_output.txt",
  classified_out = "classified.fq",
  unclassified_out = NULL,
  extract_koutput = NULL,
  extract_reads = NULL,
  extract_kreport = NULL,
  extract_kmer = NULL,
  extract_umi = NULL,
  taxon = c("d__Bacteria", "d__Fungi", "d__Viruses"),
  ranks = c("G", "S"),
  chunk_size = NULL,
  buffer_size = NULL,
  batch_size = NULL,
  nqueue = NULL,
  kmer_len = 35L,
  min_frac = 0.5,
  exclude = "9606",
  mmap_koutput = FALSE,
  mmap_reads = TRUE,
  kraken2_threads = NULL,
  kractor_threads = NULL,
  sckmer_threads = NULL,
  kraken2 = NULL,
  envpath = NULL,
  overwrite = FALSE,
  odir = getwd()
)
}
\arguments{
\item{reads}{A character vector of FASTQ files used as input to Kraken2.
Can be one file (single-end) or two files (paired-end).}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \code{kraken2} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link[=shQuote]{shQuote()}}. Details see: \code{cmd_help(kraken2())}.}

\item{ubread}{Path to the input sequence file that contains UMI and/or
barcode sequences. If \code{NULL}, UMI/barcode parsing is disabled. When
specified, \code{reads} must contain only a single FASTQ file. This is commonly
used for 10x Genomics single-cell data, where the first read file contains
only UMI and barcode sequences. In such cases, you should provide that file
via \code{ubread}.}

\item{umi_ranges}{A range or a list of ranges specifying where UMI sequences
are located in the reads. Must be created using the \code{ubrange()} function.
Only used when \code{ubread} is not \code{NULL}.}

\item{barcode_ranges}{A range or a list of ranges specifying where cell
barcode sequences are located in the reads. Must be created using the
\code{ubrange()} function.  Only used when \code{ubread} is not \code{NULL}.}

\item{barcode_extractor, umi_extractor}{A function takes sequence IDs, read1,
read2 and return a character corresponding to cell barcode and UMI
respectively, each should have the same length of the input.}

\item{db}{Path to the Kraken2 database. You can download prebuilt databases
from \url{https://benlangmead.github.io/aws-indexes/k2}, or build your own by
following the instructions at
\url{https://github.com/DerrickWood/kraken2/wiki/Manual#kraken-2-databases}.}

\item{kreport}{Path to the Kraken2 report file.}

\item{koutput}{Path to the Kraken2 output file.}

\item{classified_out}{A string of path to save classified sequences, which
should be a fastq file.}

\item{unclassified_out}{A string of path to save unclassified sequences,
which should be a fastq file.}

\item{extract_koutput}{Path to the file where the extracted Kraken2 output
matching the specified \code{taxon} will be saved. Defaults to
\code{"kraken_microbiome_output.txt"}.}

\item{extract_reads}{A character vector of the same length as \code{reads},
specifying the output FASTQ file(s) where the matching reads will be saved.
Defaults to \code{"kraken_microbiome_reads(_1|2).txt"}.}

\item{extract_kreport}{Path to the file where the extracted Kraken2 report,
matched to \code{koutput}, will be saved. Defaults to
\code{"kraken_microbiome_report.txt"}.}

\item{extract_kmer}{Path to the file where the quantified k-mer data,
derived from \code{koutput}, will be saved. Defaults to
\code{"kraken_microbiome_kmer.txt"}.}

\item{extract_umi}{Path to the file where the quantified UMI data,
derived from \code{koutput}, will be saved. Defaults to
\code{"kraken_microbiome_umi.txt"}.}

\item{taxon}{An atomic character specify the taxa name wanted. Should follow
the kraken style, connected by rank codes, two underscores, and the
scientific name of the taxon (e.g., "d__Viruses").}

\item{ranks}{Taxa ranks to analyze.}

\item{chunk_size}{Integer. Size in bytes of the intermediate chunk used to
split and distribute data to worker threads during processing. Default is
\code{10 * 1024 * 1024} (10MB).}

\item{buffer_size}{Integer specifying the buffer size in bytes used for
writing to disk. This controls the capacity of the buffered file writer.
Default is \code{1 * 1024 * 1024} (1MB).}

\item{batch_size}{Integer. Number of records to accumulate before triggering
a write operation. Default is \code{500}.}

\item{nqueue}{Integer. Maximum number of buffers per thread, controlling the
amount of in-flight data awaiting writing. Default: \code{3}.}

\item{kmer_len}{Kraken kmer length. Default: \code{35L}, which is the default kmer
size of kraken2.}

\item{min_frac}{Minimum fraction of kmers directly assigned to taxid to use
read. Reads with \verb{<=min_frac} of the k-mers map inside the taxon's lineage
are also discarded.}

\item{exclude}{A character of taxid to exclude, for \code{SAHMI}, the host taxid.
Reads with any k-mers mapped to the \code{exclude} are discarded.}

\item{mmap_koutput, mmap_reads}{Logical. Whether to enable memory-mapped file
access for reading input files (\code{mmap_reads}) and writing Kraken2 output
(\code{mmap_koutput}). When set to \code{TRUE}, the function uses memory mapping to
reduce data copying and improve performance in multi-threaded environments.
This can be highly efficient on some systems, but performance gains may
vary depending on the operating system and file system. Defaults to \code{TRUE}.}

\item{kraken2_threads}{Integer. Number of threads to use. Default to all
available threads.}

\item{kractor_threads}{Integer. Number of threads to use. Default will
determined atomatically by rayon.}

\item{sckmer_threads}{Integer. Number of threads to use. Default to all
available threads.}

\item{kraken2}{A string of path to \code{kraken2} command.}

\item{envpath}{String. Additional path to include in the \code{PATH} environment
variable. Used to locate the \code{kraken2} executable.}

\item{overwrite}{Logical. Whether to overwrite existing files in \code{odir}.}

\item{odir}{A string of path to the output directory.}
}
\value{
None. This function generates the following files:
\itemize{
\item \code{kreport}: the \code{kraken2} report file.
\item \code{koutput}: the \code{kraken2} output file.
\item \code{classified_out}: the classified sequence file(s) from \code{kraken2}. Not
required for downstream processing.
\item \code{unclassified_out}: the unclassified sequence file(s) from \code{kraken2}. Not
required for downstream processing.
\item \code{extract_koutput}: Kraken2 output entries corresponding to the specified
\code{taxon}, extracted from \code{koutput}.
\item \code{extract_reads}: Sequence file(s) containing reads assigned to the
specified \code{taxon}.
\item \code{extract_kreport}: A filtered version of the Kraken2 taxonomic report,
containing only taxa that meet the \code{ranks} criteria and are observed in
\code{koutput}. Used by \code{\link[=rpmm_quantile]{rpmm_quantile()}} and \code{\link[=slsd]{slsd()}}.
\item \code{extract_kmer}: A table quantifying total and unique k-mers assigned to
each taxon across barcodes. Used by \code{\link[=blsd]{blsd()}}.
\item \code{extract_umi}: A table of taxon–barcode–UMI combinations indicating all
observed UMI-tagged reads per taxon. Used by \code{\link[=taxa_counts]{taxa_counts()}}.
}
}
\description{
Full SAHMI-Compatible Pipeline: Kraken2 + Read Extraction + K-mer/UMI
Quantification
}
\seealso{
\itemize{
\item \code{\link[blit:kraken2]{blit::kraken2()}}
\item \code{\link[=kractor]{kractor()}}
\item \code{\link[=sckmer]{sckmer()}}
}
}
