% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{locate_path}
\alias{locate_path}
\title{Locate the output file}
\usage{
locate_path(x, sample, dir = getwd())
}
\arguments{
\item{x}{One of "kraken_report", "mpa_report", "kraken_out",
"microbiome_out", "sckmer".}

\item{sample}{A string, sample name, will be used to locate file since
\code{SAHMI} use this to create output file name.}

\item{dir}{Path to retuls directory.}
}
\value{
The path of x.
}
\description{
Locate the output file
}
