% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken_report.R
\name{parse_kraken_report}
\alias{parse_kraken_report}
\title{Parse kraken report file}
\usage{
parse_kraken_report(kraken_report, intermediate_ranks = TRUE, mpa = FALSE)
}
\arguments{
\item{kraken_report}{The path to kraken report file.}

\item{intermediate_ranks}{A bool indicates whether to include
non-traditional taxonomic ranks in output.}

\item{mpa}{A bool indicates whether to use mpa-style.}
}
\value{
A polars \link[polars:DataFrame_class]{DataFrame}.
}
\description{
Parse kraken report file
}
\seealso{
\url{https://github.com/DerrickWood/kraken2/blob/master/docs/MANUAL.markdown}
}
