% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kr.R
\name{read_kr}
\alias{read_kr}
\title{Read kraken report}
\usage{
read_kr(kraken_report, sample = NULL, min_reads = 2L, min_uniq = 2L)
}
\arguments{
\item{kraken_report}{The path to kraken2uniq report
(\emph{sample.kraken.report.txt}).}

\item{sample}{A string, sample name, will be used to create output file name.}

\item{min_reads}{Minimal number of reads.}

\item{min_uniq}{Minimal unique number.}
}
\value{
A \link[data.table:data.table]{data.table} object
}
\description{
Read kraken report
}
