% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpmm_quantile.R
\name{rpmm_quantile}
\alias{rpmm_quantile}
\title{Identifying contaminants and false positives taxa (cell line quantile test)}
\usage{
rpmm_quantile(
  kreports,
  study = "current study",
  taxon = c("d__Bacteria", "d__Fungi", "d__Viruses"),
  quantile = 0.95,
  alpha = 0.05,
  alternative = "greater",
  drop_unmatched_taxa = TRUE
)
}
\arguments{
\item{kreports}{A character of path to all kraken report files.}

\item{study}{A string of the study name, used to differentiate with cell line
data.}

\item{quantile}{Probabilities with values in \verb{[0, 1]} specifying the quantile
to calculate.}

\item{alpha}{Level of significance.}

\item{alternative}{A string specifying the alternative hypothesis, must be
one of "two.sided", "greater" (default) or "less". You can specify just the
initial letter.}

\item{drop_unmatched_taxa}{A boolean value, indicates whether taxa not found
in celllines data should be dropped. Default: \code{TRUE}.}
}
\value{
A polars \link[polars:DataFrame_class]{DataFrame} with following
attributes:
\enumerate{
\item \code{pvalues}: Quantile test pvalue.
\item \code{unmatched_taxa}: taxids in current study but not found in cellline data.
\item \code{significant}: significant taxids with \code{pvalues < alpha}.
\item \code{truly}: truly taxids based on \code{alpha} and \code{unmatched_taxa}. If
\code{drop_unmatched_taxa} isn't \code{TRUE}, this should be the union of
\code{unmatched_taxa} and \code{significant}, otherwise, this should be the same
with \code{significant}.
}
}
\description{
This function detects likely contaminant or false-positive microbial taxa
by comparing read abundance (in reads per million microbiome reads, RPMM)
from input Kraken2 reports with a curated reference of known contaminants
derived from cell lines. It performs a one-sample quantile test on each taxon
and reports taxa with significant enrichment above reference levels.
}
\examples{
\dontrun{
# `paths` should be the output directory for each sample from
# `blit::kraken2()`
quantile_test <- remove_contaminants(
    kreports = file.path(paths, "kraken_report.txt"),
    quantile = 0.99, drop_unmatched_taxa = TRUE
)
ggplot2::autoplot(quantile_test)
}
}
