% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsahmi-package.R
\docType{package}
\name{rsahmi-package}
\alias{rsahmi}
\alias{rsahmi-package}
\title{rsahmi: Single-cell Analysis of Host-Microbiome Interactions}
\description{
rsahmi enables the systematic recovery and denoising of microbial signals from genomic sequencing of host tissues.
}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunpp96@outlook.com}

}
\keyword{internal}
