% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blsd.R
\name{run_blsd}
\alias{run_blsd}
\title{Barcode level signal denoising}
\usage{
run_blsd(
  kraken_report,
  sckmer,
  min_number = 3L,
  method = "spearman",
  ...,
  p.adjust = "BH"
)
}
\arguments{
\item{kraken_report}{The path to kraken2uniq report, often in the form of
\strong{sample.kraken.report.txt}. It's easy to use \link{locate_path} to create the
path.}

\item{sckmer}{The path to sckmer results, often in the form of
\strong{sample.sckmer.txt}. It's easy to use \link{locate_path} to create the path.}

\item{min_number}{An integer, the minimal number of items per taxid}

\item{method}{A character string indicating which correlation coefficient is
to be used for the test. One of "pearson", "kendall", or "spearman", can be
abbreviated.}

\item{...}{Other arguments passed to \link[stats:cor.test]{cor.test}.}

\item{p.adjust}{Pvalue correction method, a character string. Can be
abbreviated. Details see \link[stats:p.adjust]{p.adjust}.}
}
\value{
A \link[data.table:data.table]{data.table}
}
\description{
True taxa are detected on multiple barcodes and with a proprotional number of
total and unique k-mer sequences across barcodes, measured as a significant
Spearman correlation between the number of total and unique k-mers across
barcodes.
}
