% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken2.R
\name{run_kraken2}
\alias{run_kraken2}
\title{Running Kraken2}
\usage{
run_kraken2(
  fq1,
  fq2 = NULL,
  sample = NULL,
  out_dir = getwd(),
  ncbi_blast_path = NULL,
  kraken2_db = NULL,
  mpa_report = TRUE,
  cores = availableCores(),
  cmd = NULL,
  kraken2_args = character(),
  python_cmd = NULL,
  sys_args = list()
)
}
\arguments{
\item{fq1, fq2}{Path to fastq 1 file.}

\item{sample}{A string, sample name, will be used to create output file name.}

\item{out_dir}{The path of output directory.}

\item{ncbi_blast_path}{The path of ncbi-blastx.}

\item{kraken2_db}{The path of kraken database.}

\item{mpa_report}{A scalar logical indicates whethter reporting mpa-style
results.}

\item{cores}{The number of cores to use.}

\item{cmd}{The path of \code{kraken2} command.}

\item{kraken2_args}{Other arguments passed to \code{kraken2}.}

\item{python_cmd}{The path of \code{python2} or \code{python3}.}

\item{sys_args}{Other arguments passed to \link{system2}.}
}
\description{
Metagenomic classification of paired-end reads from single-cell RNA
sequencing fastq files can be performed using any k-mer based mapper that
identifies a taxonomic ID for each k-mer and read. However, \code{SAHMI} is
optimized to run with \code{Kraken2Uniq}, which finds exact matches of candidate
35-mer genomic substrings to the lowest common ancestor of genomes in a
reference metagenomic database. It is essential that all realistically
possible genomes are included as mapping references at this stage (e.g. host,
known vectors, etc.), or that host mappable reads are excluded. The required
outputs from this step are: a Kraken summary report with sample level
metagenomic counts, a Kraken output file with read and k-mer level taxonomic
classifications, an MPA-style report, and raw sequencing fastq files with
taxonomic classification for each read. Please see
\url{https://github.com/DerrickWood/kraken2/blob/master/docs/MANUAL.markdown} for
more details on installation and usage of Kraken2/KrakenUniq.
}
\seealso{
\url{https://github.com/sjdlabgroup/SAHMI}
}
