% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kraken_microbiome.R
\name{run_kraken_microbiome}
\alias{run_kraken_microbiome}
\title{Run kraken2 and Extract Microbiome reads}
\usage{
run_kraken_microbiome(
  fq1,
  fq2 = NULL,
  sample = NULL,
  out_dir = getwd(),
  ncbi_blast_path = NULL,
  kraken2_db = NULL,
  cores = availableCores(),
  kraken2_cmd = NULL,
  kraken2_args = character(),
  python_cmd = NULL,
  microbiome_pattern = "(?i)Bacteria|Fungi|Viruses",
  ntaxid = 8000L,
  ...,
  sys_args = list()
)
}
\arguments{
\item{fq1, fq2}{The path to the classified microbiome fastq file (returned by
\code{run_kraken2}).}

\item{sample}{A string, sample name, will be used to create output file name.}

\item{out_dir}{The path of output directory.}

\item{ncbi_blast_path}{The path of ncbi-blastx.}

\item{kraken2_db}{The path of kraken database.}

\item{cores}{The number of cores to use.}

\item{kraken2_cmd}{The path of kraken2 command.}

\item{kraken2_args}{Other arguments passed to \code{kraken2}.}

\item{python_cmd}{The path of \code{python2} or \code{python3}.}

\item{microbiome_pattern}{Perl regex patterns to match the bacteria
microbiome.}

\item{ntaxid}{Number of taxids to extract at a time.}

\item{...}{Other arguments passed to \link{grepl}.}

\item{sys_args}{Other arguments passed to \link{system2}.}
}
\description{
A helper function to run both \code{run_kraken2} and \code{extract_microbiome}
}
