% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_krmikm.R
\name{run_krmikm}
\alias{run_krmikm}
\title{Run kraken2 and Extract Microbiome reads}
\usage{
run_krmikm(
  fq1,
  fq2 = NULL,
  sample = NULL,
  out_dir = getwd(),
  ncbi_blast_path = NULL,
  kraken2_db = NULL,
  cores = availableCores(),
  kraken2_cmd = NULL,
  kraken2_args = character(),
  python_cmd = NULL,
  microbiome_pattern = "(?i)Bacteria|Fungi|Viruses",
  ntaxid = 8000L,
  ranks = c("G", "S"),
  cb_len = 16L,
  umi_len = 10L,
  host = 9606L,
  nsample = 1000L,
  kmer_len = 35L,
  min_frac = 0.5,
  ...,
  sys_args = list()
)
}
\arguments{
\item{fq1, fq2}{The path to the classified microbiome fastq file (returned by
\code{run_kraken2}).}

\item{sample}{A string, sample name, will be used to create output file name.}

\item{out_dir}{The path of output directory.}

\item{ncbi_blast_path}{The path of ncbi-blastx.}

\item{kraken2_db}{The path of kraken database.}

\item{cores}{The number of cores to use.}

\item{kraken2_cmd}{The path of kraken2 command.}

\item{kraken2_args}{Other arguments passed to \code{kraken2}.}

\item{python_cmd}{The path of \code{python2} or \code{python3}.}

\item{microbiome_pattern}{Perl regex patterns to match the bacteria
microbiome.}

\item{ntaxid}{Number of taxids to extract at a time.}

\item{ranks}{Taxa ranks to analyze.}

\item{cb_len}{Nucleutide length of cell barcodes}

\item{umi_len}{Nucleutide length of umis}

\item{host}{Host taxid to exclude}

\item{nsample}{Max number of barcodes to sample per taxa.}

\item{kmer_len}{Kraken kmer length.}

\item{min_frac}{Minimum fraction of kmers directly assigned to taxid to use
read.}

\item{...}{Other arguments passed to \link{grepl}.}

\item{sys_args}{Other arguments passed to \link{system2}.}
}
\description{
A helper function to run \code{run_kraken2}, \code{extract_microbiome} and \code{run_sckmer}
sequentially.
}
