% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sckmer.R
\name{run_sckmer}
\alias{run_sckmer}
\title{Tabulating k-mer statistics}
\usage{
run_sckmer(
  fa1,
  fa2 = NULL,
  kraken_report = NULL,
  mpa_report = NULL,
  microbiome_out = NULL,
  sample = NULL,
  out_dir = getwd(),
  ranks = c("G", "S"),
  cb_len = 16L,
  umi_len = 10L,
  host = 9606L,
  nsample = 1000L,
  kmer_len = 35L,
  min_frac = 0.5,
  cores = availableCores()
)
}
\arguments{
\item{fa1, fa2}{The path to microbiome fasta 1 and 2 file (returned by
\code{extract_microbiome}).}

\item{kraken_report}{The path to kraken2uniq report, if \code{NULL},
\strong{sample.kraken.report.txt} in \code{out_dir}.}

\item{mpa_report}{A scalar logical indicates whethter reporting mpa-style
results.}

\item{microbiome_out}{The path of microbiome output file (returned by
\link{extract_microbiome}). if \code{NULL}, will locate
\strong{sample.microbiome.output.txt} in \code{out_dir}.}

\item{sample}{A string, sample name, will be used to create output file name.}

\item{out_dir}{The path of output directory.}

\item{ranks}{Taxa ranks to analyze.}

\item{cb_len}{Nucleutide length of cell barcodes}

\item{umi_len}{Nucleutide length of umis}

\item{host}{Host taxid to exclude}

\item{nsample}{Max number of barcodes to sample per taxa.}

\item{kmer_len}{Kraken kmer length.}

\item{min_frac}{Minimum fraction of kmers directly assigned to taxid to use
read.}

\item{cores}{The number of cores to use.}
}
\description{
The function count the number of k-mers and unique k-mers
assigned to a taxon across barcodes. The cell barcode and unique molecular
identifier (UMI) are used to identify unique barcodes and reads. Data is
reported for taxa of pre-specified ranks (default genus + species) taking
into account all subsequently higher resolution ranks. Reads with any k-mers
mapped to the host (e.g. human) are discarded. Reads with >50\% of the k-mers
map outside the taxon's lineage are also discarded. The output is a table of
barcodes, taxonomic IDs, number of k-mers, and number of unique k-mers.
}
\seealso{
\url{https://github.com/sjdlabgroup/SAHMI}
}
