% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slsd.R
\name{sahmi_slsd}
\alias{sahmi_slsd}
\title{Sample level signal denoising}
\usage{
sahmi_slsd(
  kreports,
  method = "spearman",
  ...,
  min_reads = 2L,
  min_uniq = 2L,
  min_number = 3L
)
}
\arguments{
\item{kraken_reports}{The paths to kraken2uniq report for all samples, often
in the form of \strong{sample.kraken.report.txt}. It's easy to use \link{locate_path}
to create the path.}

\item{samples}{An atomic character specifying sample names.}
}
\value{
A \link[data.table:data.table]{data.table} of correlation coefficient and
pvalue for \verb{cor(#k-mers, #unique k-mers)} (r1 and p1), \verb{cor(#k-mers, #reads)} (r2 and p2) and \verb{cor(#reads, #unique k-mers)} (r3 and p3).
}
\description{
In the low-microbiome biomass setting, real microbes also exhibit a
proportional number of total k-mers, number of unique k-mers, as well as
number of total assigned sequencing reads across samples; i.e. the following
three Spearman correlations are significant when tested using sample-level
data provided in Kraken reports: \verb{cor(#k-mers, #unique k-mers)},
\verb{cor(#k-mers, #reads)} and \verb{cor(#reads, #unique k-mers)}.
}
