% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sckmer-polars.R
\name{sckmer_polars}
\alias{sckmer_polars}
\title{Tabulating k-mer statistics}
\usage{
sckmer_polars(
  fa1,
  kraken_report,
  kraken_out,
  fa2 = NULL,
  cb_and_umi = function(sequence_id, read1, read2) {
     lapply(strsplit(sequence_id,
    "_", fixed = TRUE), `[`, 2:3)
 },
  ranks = c("G", "S"),
  kmer_len = 35L,
  min_frac = 0.5,
  exclude = "9606",
  threads = 10L
)
}
\arguments{
\item{fa1, fa2}{The path to microbiome fasta 1 and 2 file (returned by
\code{extract_kraken_reads}).}

\item{kraken_report}{The path to kraken report file.}

\item{kraken_out}{The path of microbiome output file. Usually should be
filtered with \link{extract_kraken_output}.}

\item{cb_and_umi}{A function takes sequence id, read1, read2 and return a
list of the same of the input, each have two elements correspond to cell
barcode and UMI respectively.}

\item{ranks}{Taxa ranks to analyze.}

\item{kmer_len}{Kraken kmer length. Default: \code{35L}, which is the default kmer
size of kraken2.}

\item{min_frac}{Minimum fraction of kmers directly assigned to taxid to use
read. Reads with \verb{<=min_frac} of the k-mers map inside the taxon's lineage
are also discarded.}

\item{exclude}{A character of taxid to exclude, for \code{SAHMI}, the host taxid.
Reads with any k-mers mapped to the \code{exclude} are discarded.}
}
\description{
The function count the number of k-mers and unique k-mers
assigned to a taxon across barcodes. The cell barcode and unique molecular
identifier (UMI) are used to identify unique barcodes and reads. Data is
reported for taxa of pre-specified ranks (default genus + species) taking
into account all subsequently higher resolution ranks. The output is a table
of barcodes, taxonomic IDs, number of k-mers, and number of unique k-mers.
}
\seealso{
\url{https://github.com/sjdlabgroup/SAHMI}
}
