% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubrange.R
\name{ubrange}
\alias{ubrange}
\title{Create a UMI/Barcode range}
\usage{
ubrange(start = NULL, end = NULL)
}
\arguments{
\item{start}{Integer. Start position (1-based). Optional if \code{end} is
provided.}

\item{end}{Integer. End position (1-based). Optional if \code{start} is provided.}
}
\value{
A \code{rsahmi_ubrange} object.
}
\description{
Constructs a \code{rsahmi_ubrange} object representing a UMI/Barcode range with
optional start and end positions.
}
\examples{
# Create a single range
ubrange(start = 100, end = 200)

# If `end` is NULL, the range goes to the end of the sequence
ubrange(start = 100)

# If `start` is NULL, the range starts from the beginning of the sequence
ubrange(end = 200)

# Combine multiple ranges
r1 <- ubrange(100, 200)
r2 <- ubrange(300, 400)
r3 <- ubrange(end = 500)

# Combine them into a multi-range object using `c()`
c(r1, r2, r3)

# Nested combinations are also supported
c(c(r1, r2), r3)
}
