% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bench_read}
\alias{bench_read}
\title{Benchmark File Reading with Optional Memory Mapping}
\usage{
bench_read(file, chunk_size = NULL, mmap = TRUE)
}
\arguments{
\item{chunk_size}{Integer. Size in bytes of the intermediate chunk used to
split and distribute data to worker threads during processing. Default is
\code{10 * 1024 * 1024} (10MB).}
}
\description{
This function benchmarks the reading performance of a file by processing it
in chunks. It allows for optional memory-mapped file access, which can
improve performance in multi-threaded environments by reducing memory
copying.
}
