% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blsd.R
\name{blsd}
\alias{blsd}
\title{Barcode level signal denoising}
\usage{
blsd(
  kmer,
  method = "spearman",
  ...,
  p.adjust = "BH",
  min_kmer_len = 3L,
  min_number = 3L
)
}
\arguments{
\item{kmer}{kmer data returned by \link{kmer}.}

\item{method}{A character string indicating which correlation coefficient is
to be used for the test. One of "pearson", "kendall", or "spearman", can be
abbreviated.}

\item{...}{Other arguments passed to \link[stats:cor.test]{cor.test}.}

\item{p.adjust}{Pvalue correction method, a character string. Can be
abbreviated. Details see \link[stats:p.adjust]{p.adjust}.}

\item{min_kmer_len}{An integer, the minimal number of kmer to filter taxa.
SAHMI use \code{2}.}

\item{min_number}{An integer, the minimal number of cell per taxid. SAHMI use
\code{4}.}
}
\description{
True taxa are detected on multiple barcodes and with a proprotional number of
total and unique k-mer sequences across barcodes, measured as a significant
Spearman correlation between the number of total and unique k-mers across
barcodes. (\code{padj < 0.05})
}
\seealso{
https://github.com/sjdlabgroup/SAHMI
}
