% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contaminants.R
\name{contaminants}
\alias{contaminants}
\title{Identifying contaminants and false positives (cell line quantile test)}
\usage{
contaminants(
  kraken_report,
  taxon = c("d__Bacteria", "d__Fungi", "d__Viruses"),
  quantile = 0.99
)
}
\arguments{
\item{kraken_report}{The path to kraken report file.}

\item{taxon}{An atomic character specify the taxa name wanted. Should follow
the kraken style, connected by rank codes, two underscores, and the
scientific name of the taxon (e.g., "d__Viruses")}

\item{quantile}{Probabilities with values in \verb{[0, 1]} specifying the quantile
to calculate.}
}
\value{
A polars \link[polars:DataFrame_class]{DataFrame}
}
\description{
Identifying contaminants and false positives (cell line quantile test)
}
