% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor.R
\name{extractor}
\alias{extractor}
\alias{extract_taxids}
\alias{extract_kraken_output}
\alias{extract_kraken_reads}
\title{Extract reads and output from Kraken}
\usage{
extract_taxids(
  kraken_report,
  taxon = c("d__Bacteria", "d__Fungi", "d__Viruses")
)

extract_kraken_output(
  kraken_out,
  taxids,
  ofile = "kraken_microbiome_output.txt",
  odir = getwd(),
  ...
)

extract_kraken_reads(
  kraken_out,
  reads,
  ofile = NULL,
  odir = getwd(),
  threads = NULL,
  ...,
  seqkit = NULL
)
}
\arguments{
\item{kraken_report}{The path to kraken report file.}

\item{taxon}{An atomic character specify the taxa name wanted. Should follow
the kraken style, connected by rank codes, two underscores, and the
scientific name of the taxon (e.g., "d__Viruses")}

\item{kraken_out}{The path to kraken output file.}

\item{taxids}{A character specify taxa identifier to extract.}

\item{ofile}{A string of file save the kraken output of specified \code{taxids}.}

\item{odir}{A string of directory to save the \code{ofile}.}

\item{...}{\itemize{
\item \code{extract_kraken_output}: Additional arguments passed to
\link[polars:IO_sink_csv]{sink_csv}.
\item \code{extract_kraken_reads}: Additional arguments passed to
\link[biosys:Execute]{run} method.
}}

\item{reads}{The original fastq files (input in \code{kraken2}). You can pass
two paired-end files directly.}

\item{threads}{Number of threads to use, see \code{biosys::seqkit("grep")$help()}.}

\item{seqkit}{A string of path to \code{seqkit} command.}
}
\description{
Extract reads and output from Kraken
}
\seealso{
\url{https://github.com/DerrickWood/kraken2/blob/master/docs/MANUAL.markdown}
}
