% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kractor.R
\name{kractor}
\alias{kractor}
\title{Extract Reads and Output from Kraken2 (kraken2 extractor)}
\usage{
kractor(
  kreport,
  koutput,
  reads,
  extract_koutput = NULL,
  extract_reads = NULL,
  taxon = c("d__Bacteria", "d__Fungi", "d__Viruses"),
  buffer_size = NULL,
  odir = getwd()
)
}
\arguments{
\item{kreport}{Path to the Kraken2 report file.}

\item{koutput}{Path to the Kraken2 output file.}

\item{reads}{A character vector of FASTQ files used as input to Kraken2.
Can be one file (single-end) or two files (paired-end).}

\item{extract_koutput}{Path to the file where the extracted Kraken2 output
matching the specified \code{taxon} will be saved. Defaults to
\code{"kraken_microbiome_output.txt"}.}

\item{extract_reads}{A character vector of the same length as \code{reads},
specifying the output FASTQ file(s) where the matching reads will be saved.
Defaults to \code{"kraken_microbiome_reads(_1|2).txt"}}

\item{taxon}{An atomic character specify the taxa name wanted. Should follow
the kraken style, connected by rank codes, two underscores, and the
scientific name of the taxon (e.g., "d__Viruses").}

\item{buffer_size}{Integer specifying the buffer size in bytes for reading
and writing files. Note that more than twice the specified \code{buffer_size} will
be used internally.}

\item{odir}{A string of directory to save the \code{ofile}.}
}
\description{
This function extracts reads and classification results from Kraken2 output,
based on a set of specified taxonomic IDs.
}
\examples{
\dontrun{
# For 10x Genomic data, `fq1` only contain barcode and umi, but the official
# didn't give any information for this. In this way, I prefer using
# `umi-tools` to transform the `umi` into fq2 and then run `rsahmi` with
# only fq2.
blit::kraken2(
    fq1 = fq1,
    fq2 = fq2,
    classified_out = "classified.fq",
    # Number of threads to use
    blit::arg("--threads", 10L, format = "\%d"),
    # the kraken database
    blit::arg("--db", kraken_db),
    "--use-names", "--report-minimizer-data",
) |> blit::cmd_run()

# 1. `kreport` should be the kraken2 report file of `blit::kraken2()`
# 2. `koutput` should be the kraken2 output file of `blit::kraken2()`
# 3. `reads` should be the same with `fq1` and `fq2` in `blit::kraken2()`
kractor(
    kreport = "kraken_report.txt",
    koutput = "kraken_output.txt",
    reads = c(fq1, fq2)
)
}
}
\seealso{
\url{https://github.com/DerrickWood/kraken2/blob/master/docs/MANUAL.markdown}
}
