% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kractor.R
\name{kractor}
\alias{kractor}
\title{Extract Reads and Output from Kraken2 (kraken2 extractor)}
\usage{
kractor(
  kreport,
  koutput,
  reads,
  extract_koutput = NULL,
  extract_reads = NULL,
  taxon = c("d__Bacteria", "d__Fungi", "d__Viruses"),
  read_buffer = NULL,
  write_buffer = NULL,
  batch_size = NULL,
  read_queue = NULL,
  write_queue = NULL,
  threads = NULL,
  odir = getwd()
)
}
\arguments{
\item{kreport}{Path to the Kraken2 report file.}

\item{koutput}{Path to the Kraken2 output file.}

\item{reads}{A character vector of FASTQ files used as input to Kraken2.
Can be one file (single-end) or two files (paired-end).}

\item{extract_koutput}{Path to the file where the extracted Kraken2 output
matching the specified \code{taxon} will be saved. Defaults to
\code{"kraken_microbiome_output.txt"}.}

\item{extract_reads}{A character vector of the same length as \code{reads},
specifying the output FASTQ file(s) where the matching reads will be saved.
Defaults to \code{"kraken_microbiome_reads(_1|2).txt"}}

\item{taxon}{An atomic character specify the taxa name wanted. Should follow
the kraken style, connected by rank codes, two underscores, and the
scientific name of the taxon (e.g., "d__Viruses").}

\item{read_buffer}{Integer specifying the size in bytes of the intermediate
buffer used for splitting and distributing chunks to worker threads during
processing. Default is \code{2 * 1024 * 1024} (1MB).}

\item{write_buffer}{Integer specifying the buffer size in bytes used for
writing to disk. This controls the capacity of the buffered file writer.
Default is \code{1 * 1024 * 1024} (2MB).}

\item{batch_size}{Integer. Number of records to accumulate before triggering
a write operation. Default is \code{200}.}

\item{read_queue, write_queue}{Integer. Maximum number of buffers per thread,
controlling the amount of in-flight data awaiting processing or
writing.}

\item{threads}{Integer. Number of threads to use. Default is \code{1}. Performance
is generally constrained by system call I/O, so increasing thread count may
not substantially reduce runtime.}

\item{odir}{A string of directory to save the \code{ofile}.}
}
\value{
None. This function generates the following files:
\itemize{
\item \code{extract_koutput}: Kraken2 output entries corresponding to the specified
\code{taxon}, extracted from koutput.
\item \code{extract_reads}: Sequence file(s) containing reads assigned to the
specified \code{taxon}.
}
}
\description{
This function extracts reads and classification results from Kraken2 output,
based on a set of specified taxonomic IDs.
}
\examples{
\dontrun{
# For 10x Genomic data, `fq1` only contain barcode and umi, but the official
# didn't give any information for this. In this way, I prefer using
# `umi-tools` to transform the `umi` into fq2 and then run `rsahmi` with
# only fq2.
blit::kraken2(
    fq1 = fq1,
    fq2 = fq2,
    classified_out = "classified.fq",
    # Number of threads to use
    blit::arg("--threads", 10L, format = "\%d"),
    # the kraken database
    blit::arg("--db", kraken_db),
    "--use-names", "--report-minimizer-data",
) |> blit::cmd_run()

# 1. `kreport` should be the kraken2 report file of `blit::kraken2()`
# 2. `koutput` should be the kraken2 output file of `blit::kraken2()`
# 3. `reads` should be the same with `fq1` and `fq2` in `blit::kraken2()`
kractor(
    kreport = "kraken_report.txt",
    koutput = "kraken_output.txt",
    reads = c(fq1, fq2)
)
}
}
\seealso{
\code{\link[=kraken_taxon]{kraken_taxon()}}
}
