% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kractor.R
\name{kractor_koutput}
\alias{kractor_koutput}
\title{Filter Kraken2 Output by Taxon}
\usage{
kractor_koutput(
  kreport,
  koutput,
  extract_koutput = NULL,
  taxonomy = c("d__Bacteria", "d__Fungi", "d__Viruses"),
  ranks = c("G", "S"),
  taxa = NULL,
  taxids = NULL,
  descendants = TRUE,
  chunk_size = NULL,
  buffer_size = NULL,
  batch_size = NULL,
  nqueue = NULL,
  mmap = TRUE,
  threads = NULL,
  odir = NULL
)
}
\arguments{
\item{kreport}{Path to the Kraken2 report file.}

\item{koutput}{Path to the Kraken2 output file.}

\item{extract_koutput}{Path to the file where the extracted Kraken2 output
matching the specified \code{taxon} will be saved. Defaults to
\code{"kraken_microbiome_output.txt"}.}

\item{taxonomy}{Character vector. The set of taxonomic groups to include
(default: \code{c("d__Bacteria", "d__Fungi", "d__Viruses")}). This defines the
global taxa to consider. If \code{NULL}, all taxa will be used. If \code{descendants = TRUE}, only the descendants within these groups will be considered. The
selection of taxa can be further refined using the \code{ranks}, \code{taxa}, and
\code{taxids} parameters. One of \code{taxonomy}, \code{ranks}, \code{taxa}, or \code{taxids} must be
provided.}

\item{ranks}{Character vector. The taxonomic ranks to filter by. Default:
\code{c("G", "S")} (optional).}

\item{taxa}{Character vector. Specific taxa to include (optional).}

\item{taxids}{Character vector. A list of taxid values to filter by
(optional).}

\item{descendants}{Logical. Whether to include descendants of the selected
taxa (default: \code{TRUE}).}

\item{batch_size}{Integer. Number of FASTQ records to accumulate before
dispatching a chunk to worker threads for processing. This controls the
granularity of parallel work and affects memory usage and performance.
Default is \code{256}.}

\item{nqueue}{Integer. Maximum number of buffers per thread, controlling the
amount of in-flight data awaiting writing. Default: \code{3}.}

\item{mmap}{Logical. Whether to enable memory-mapped file access for reading
input Kraken2 output. When set to \code{TRUE}, the function uses memory mapping
to reduce data copying and improve performance in multi-threaded
environments. This can be highly efficient on some systems, but
performance gains may vary depending on the operating system and file
system. Defaults to \code{TRUE}.}

\item{threads}{Integer. Number of threads to use. Default: \code{3}.}

\item{odir}{A string of directory to save the output files. Please see
\code{Value} section for details.}
}
\value{
None. The function generates a filtered Kraken2 output file
containing entries corresponding to the specified \code{taxonomy}, \code{ranks},
\code{taxa}, \code{taxids}, and \code{descendants} extracted from the input \code{koutput}.
}
\description{
This function filters Kraken2 classification output (\code{koutput}) by taxonomic
identifiers derived from a Kraken2 report (\code{kreport}). It extracts lines
matching the desired \code{taxonomy}, \code{ranks}, \code{taxa}, \code{taxids}, and \code{descendants}
and writes the filtered results to an output file.
}
