% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kractor.R
\name{kractor_reads}
\alias{kractor_reads}
\title{Extract Reads from Kraken2 Output Based on Classification}
\usage{
kractor_reads(
  koutput,
  reads,
  extract_reads = NULL,
  chunk_size = NULL,
  buffer_size = NULL,
  batch_size = NULL,
  nqueue = NULL,
  threads = NULL,
  odir = NULL,
  mmap = TRUE
)
}
\arguments{
\item{koutput}{Path to the Kraken2 output file.}

\item{reads}{A character vector of FASTQ file paths, either the original
reads used as input to Kraken2 or the classified output reads (recommended
for efficiency as they are smaller). Accepts one file for single-end or two
files for paired-end. Note: if \code{ubread} is specified, \code{reads} must be the
original Kraken2 input reads to ensure correct barcode/UMI pairing.}

\item{extract_reads}{A character vector of the same length as \code{reads},
specifying the output FASTQ file(s) where the matching reads will be saved.
Defaults to \code{"kraken_microbiome_reads(_1|2).txt"}.}

\item{chunk_size}{Integer. Size in bytes of the intermediate chunk used to
split and distribute data to worker threads during processing. Default is
\code{10 * 1024 * 1024} (10MB).}

\item{buffer_size}{Integer specifying the buffer size in bytes used for
writing to disk. This controls the capacity of the buffered file writer.
Default is \code{1 * 1024 * 1024} (1MB).}

\item{batch_size}{Integer. Number of records to accumulate before triggering
a write operation. Default is \code{500}.}

\item{nqueue}{Integer. Maximum number of buffers per thread, controlling the
amount of in-flight data awaiting writing. Default: \code{3}.}

\item{threads}{Integer. Number of threads to use. Default will use all
available threads.}

\item{odir}{A string of directory to save the output files. Please see
\code{Value} section for details.}

\item{mmap}{Logical. Whether to enable memory-mapped file access for reading
input sequence files. When set to \code{TRUE}, the function uses memory mapping
to reduce data copying and improve performance in multi-threaded
environments. This can be highly efficient on some systems, but
performance gains may vary depending on the operating system and file
system. Defaults to \code{TRUE}.}
}
\description{
This function extracts reads corresponding to selected classifications from a
Kraken2 output file (\code{koutput}). Only reads classified to selected taxa will
be extracted from the provided sequence file (\code{reads}).
}
