% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken_taxon.R
\name{kraken_taxon}
\alias{kraken_taxon}
\title{Extract microbiome Kraken2 output and sequence reads}
\usage{
kraken_taxon(
  fq1,
  ...,
  fq2 = NULL,
  db = NULL,
  kreport = "kraken_report.txt",
  koutput = "kraken_output.txt",
  classified_out = "classified.fq",
  unclassified_out = NULL,
  extract_koutput = NULL,
  extract_reads = NULL,
  taxon = c("d__Bacteria", "d__Fungi", "d__Viruses"),
  buffer_size = NULL,
  queue_capacity = NULL,
  threads = NULL,
  kraken2 = NULL,
  envpath = NULL,
  overwrite = FALSE,
  odir = getwd()
)
}
\arguments{
\item{fq1, fq2}{A string of fastq file path.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \code{kraken2} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link[=shQuote]{shQuote()}}. Details see: \code{cmd_help(kraken2())}.}

\item{db}{Path to the Kraken2 database. You can download prebuilt databases
from \url{https://benlangmead.github.io/aws-indexes/k2}, or build your own by
following the instructions at
\url{https://github.com/DerrickWood/kraken2/wiki/Manual#kraken-2-databases}.}

\item{kreport}{Path to the Kraken2 report file.}

\item{koutput}{Path to the Kraken2 output file.}

\item{classified_out}{A string of path to save classified sequences, which
should be a fastq file.}

\item{unclassified_out}{A string of path to save unclassified sequences,
which should be a fastq file.}

\item{extract_koutput}{Path to the file where the extracted Kraken2 output
matching the specified \code{taxon} will be saved. Defaults to
\code{"kraken_microbiome_output.txt"}.}

\item{extract_reads}{A character vector of the same length as \code{reads},
specifying the output FASTQ file(s) where the matching reads will be saved.
Defaults to \code{"kraken_microbiome_reads(_1|2).txt"}}

\item{taxon}{An atomic character specify the taxa name wanted. Should follow
the kraken style, connected by rank codes, two underscores, and the
scientific name of the taxon (e.g., "d__Viruses").}

\item{buffer_size}{Integer specifying the size in bytes of the intermediate
buffer used for splitting and distributing chunks to worker threads during
processing.}

\item{queue_capacity}{Integer. Maximum number of buffers per thread,
controlling the amount of in-flight data awaiting processing or writing.
Higher values can improve throughput but increase memory use. Default is
\code{100}.}

\item{threads}{Integer. Number of threads to use. Defaults to all available
threads.}

\item{kraken2}{A string of path to \code{kraken2} command.}

\item{envpath}{String. Additional path to include in the \code{PATH} environment
variable. Used to locate the \code{kraken2} executable.}

\item{overwrite}{Logical. Whether to overwrite existing files in \code{odir}.}

\item{odir}{A string of path to the output directory.}
}
\value{
None. This function generates the following files:
\itemize{
\item \code{kreport}: the \code{kraken2} report file.
\item \code{koutput}: the \code{kraken2} output file.
\item \code{classified_out}: the classified sequence file(s) from \code{kraken2}.
\item \code{unclassified_out}: the unclassified sequence file(s) from \code{kraken2}.
\item \code{extract_koutput}: Kraken2 output entries corresponding to the specified
\code{taxon}, extracted from koutput.
\item \code{extract_reads}: Sequence file(s) containing reads assigned to the
specified \code{taxon}.
}
}
\description{
Extract microbiome Kraken2 output and sequence reads
}
\seealso{
\code{\link[=kractor]{kractor()}}
}
