% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken_taxon.R
\name{kraken_taxon}
\alias{kraken_taxon}
\title{Extracting microbiome kraken2 output and sequence reads}
\usage{
kraken_taxon(
  fq1,
  ...,
  fq2 = NULL,
  db = NULL,
  kreport = "kraken_report.txt",
  koutput = "kraken_output.txt",
  classified_out = "classified.fq",
  unclassified_out = NULL,
  extract_koutput = NULL,
  extract_reads = NULL,
  taxon = c("d__Bacteria", "d__Fungi", "d__Viruses"),
  buffer_size = NULL,
  threads = NULL,
  odir = getwd(),
  kraken2 = NULL,
  envpath = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{fq1, fq2}{A string of fastq file path.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \code{kraken2} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link[=shQuote]{shQuote()}}. Details see: \code{cmd_help(kraken2())}.}

\item{kreport}{Path to the Kraken2 report file.}

\item{koutput}{Path to the Kraken2 output file.}

\item{classified_out}{A string of path to save classified sequences, which
should be a fastq file.}

\item{unclassified_out}{A string of path to save unclassified sequences,
which should be a fastq file.}

\item{extract_koutput}{Path to the file where the extracted Kraken2 output
matching the specified \code{taxon} will be saved. Defaults to
\code{"kraken_microbiome_output.txt"}.}

\item{extract_reads}{A character vector of the same length as \code{reads},
specifying the output FASTQ file(s) where the matching reads will be saved.
Defaults to \code{"kraken_microbiome_reads(_1|2).txt"}}

\item{taxon}{An atomic character specify the taxa name wanted. Should follow
the kraken style, connected by rank codes, two underscores, and the
scientific name of the taxon (e.g., "d__Viruses").}

\item{buffer_size}{Integer specifying the buffer size in bytes for reading
and writing files. Note that more than twice the specified \code{buffer_size} will
be used internally.}

\item{odir}{A string of path to the output directory.}

\item{kraken2}{A string of path to \code{kraken2} command.}
}
\description{
Extracting microbiome kraken2 output and sequence reads
}
