% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken_report.R
\name{read_kreport}
\alias{read_kreport}
\title{Parse kraken report file}
\usage{
read_kreport(kreport, taxonomy = NULL)
}
\arguments{
\item{kreport}{The path to kraken report file.}

\item{taxonomy}{A character vector. The set of taxonomic groups to include.}
}
\value{
A data frame.
}
\description{
Parse kraken report file
}
\seealso{
\url{https://github.com/DerrickWood/kraken2/blob/master/docs/MANUAL.markdown}
}
