% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contaminants.R
\name{remove_contaminants}
\alias{remove_contaminants}
\title{Identifying contaminants and false positives taxa (cell line quantile test)}
\usage{
remove_contaminants(
  kraken_reports,
  study = "current study",
  taxon = c("d__Bacteria", "d__Fungi", "d__Viruses"),
  quantile = 0.95,
  alpha = 0.05,
  alternative = "greater"
)
}
\arguments{
\item{kraken_reports}{A character of path to all kraken report files.}

\item{study}{A string of the study name, used to differentiate with cell line
data.}

\item{taxon}{An atomic character specify the taxa name wanted. Should follow
the kraken style, connected by rank codes, two underscores, and the
scientific name of the taxon (e.g., "d__Viruses")}

\item{quantile}{Probabilities with values in \verb{[0, 1]} specifying the quantile
to calculate.}

\item{alpha}{Level of significance.}

\item{alternative}{A string specifying the alternative hypothesis, must be
one of "two.sided", "greater" (default) or "less". You can specify just the
initial letter.}
}
\value{
A polars \link[polars:DataFrame_class]{DataFrame} with following
attributes:
\enumerate{
\item \code{pvalues}: Quantile test pvalue.
\item \code{truly}: significant taxids based on \code{alpha}.
}
}
\description{
Identifying contaminants and false positives taxa (cell line quantile test)
}
