% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsahmi-package.R
\docType{package}
\name{rsahmi-package}
\alias{rsahmi}
\alias{rsahmi-package}
\title{rsahmi: Single-Cell Analysis of Host-Microbiome Interactions}
\description{
A computational resource designed to accurately detect microbial nucleic acids while filtering out contaminants and false-positive taxonomic assignments from standard transcriptomic sequencing of mammalian tissues. For more details, see Ghaddar (2023) \doi{10.1038/s43588-023-00507-1}. This implementation leverages the 'polars' package for fast and systematic microbial signal recovery and denoising from host tissue genomic sequencing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Yunuuuu/rsahmi}
  \item Report bugs at \url{https://github.com/Yunuuuu/rsahmi/issues}
}

}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunp96@163.com} (\href{https://orcid.org/0000-0003-2801-3332}{ORCID})

}
\keyword{internal}
