% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq-action.R
\name{subseq_actions}
\alias{subseq_actions}
\alias{embed}
\alias{trim}
\alias{embed_trim}
\title{Define sequence range behavior: embed, trim, or both}
\usage{
embed(ranges, tag = NULL)

trim(ranges)

embed_trim(ranges, tag = NULL)
}
\arguments{
\item{ranges}{A range or a list of ranges specifying the subsequence(s) to
process. Must be created using the \code{\link[=seq_range]{seq_range()}} function.}

\item{tag}{An optional character label to use when embedding in the header
(for \code{embed()} and \code{embed_trim()}).}
}
\value{
An annotated \code{rsahmi_seq_range} or \code{rsahmi_seq_ranges} object. object
with behavior-specific class:
\itemize{
\item \code{rsahmi_embed} — for embedding only
\item \code{rsahmi_trim} — for trimming only
\item \code{rsahmi_embed_trim} — for embedding and trimming
}
}
\description{
These functions annotate a range or a list of ranges to indicate how
the specified subsequences (e.g. UMI or barcode) should be handled during
FASTQ processing.
}
\details{
\itemize{
\item \code{embed()} extracts the sequence and appends it to the read header without
modifying the sequence.
\item \code{trim()} removes the specified subsequence from the read but does not store
it.
\item \code{embed_trim()} both extracts the subsequence for the header and removes it
from the read.
}

Each function wraps the input range object in a new class to indicate its
behavior downstream.
}
