% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_counts.R
\name{taxa_counts}
\alias{taxa_counts}
\title{Quantitation of microbes}
\usage{
taxa_counts(umi, taxids = NULL)
}
\arguments{
\item{umi}{UMI data returned by \link{prep_dataset}.}

\item{taxids}{Taxa ids to be extracted, if \code{NULL}, all taxon in \code{umi} will be
counting.}
}
\description{
After identifying true taxa, reads assigned to those taxa are
extracted and then undergo a series of filters. The cell barcode and UMI are
used to demultiplex the reads and create a barcode x taxa counts matrix. The
full taxonomic classification of all resulting barcodes and the number of
counts assigned to each clade are tabulated.
}
\seealso{
\url{https://github.com/sjdlabgroup/SAHMI}
}
