% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_counts.R
\name{taxa_counts}
\alias{taxa_counts}
\title{Quantitation of microbes}
\usage{
taxa_counts(
  fa1,
  fa2,
  kraken_report,
  mpa_report,
  taxa,
  cb_len = 16L,
  umi_len = 10L,
  n_filter = 130L,
  cores = availableCores()
)
}
\arguments{
\item{fa1, fa2}{The path to microbiome fasta 1 and 2 file (returned by
\code{extract_microbiome}).}

\item{kraken_report}{The path to kraken2uniq report
(\emph{sample.kraken.report.txt}).}

\item{mpa_report}{A scalar logical indicates whethter reporting mpa-style
results.}

\item{cb_len}{Nucleutide length of cell barcodes}

\item{umi_len}{Nucleutide length of umis}

\item{cores}{The number of cores to use.}
}
\description{
After identifying true taxa, reads assigned to those taxa are
extracted and then undergo a series of filters. The cell barcode and UMI are
used to demultiplex the reads and create a barcode x taxa counts matrix. The
full taxonomic classification of all resulting barcodes and the number of
counts assigned to each clade are tabulated.
}
